+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2024-01-11
+
+### Added
+
+- Added missing properties to OpenAI class (apiKey, organization, origin)
+- Fixed spelling of "organization"
+- Sensitive Parameter value redaction (API Key)
+
 ## [2.1.0] - 2023-11-06
 
 ### Added